+++ e3nn_jax/__init__.py
@@ -52,7 +52,19 @@ from e3nn_jax._src.su2 import su2_clebsch_gordan, su2_generators
 from e3nn_jax._src.so3 import clebsch_gordan, generators
 from e3nn_jax._src.irreps import Irrep, MulIrrep, Irreps
 from e3nn_jax._src.irreps_array import IrrepsArray
-from e3nn_jax._src.basic import concatenate, stack, mean, norm, normal, dot, cross
+from e3nn_jax._src.basic import (
+    from_chunks,
+    as_irreps_array,
+    zeros,
+    zeros_like,
+    concatenate,
+    stack,
+    mean,
+    norm,
+    normal,
+    dot,
+    cross,
+)
 from e3nn_jax._src.basic import sum_ as sum
 from e3nn_jax._src.spherical_harmonics import spherical_harmonics, sh, legendre
 from e3nn_jax._src.radial import (
@@ -158,6 +170,10 @@ __all__ = [
     "MulIrrep",  # not in docs
     "Irreps",
     "IrrepsArray",
+    "from_chunks",
+    "as_irreps_array",
+    "zeros",
+    "zeros_like",
     "concatenate",
     "stack",
     "mean",