+++ e3nn_jax/_src/dropout_haiku.py
@@ -51,25 +51,28 @@ class Dropout(hk.Module):
             return x
 
         if self.irreps is not None:
-            x = x._convert(self.irreps)
+            x = x.rechunk(self.irreps)
         if not isinstance(x, IrrepsArray):
             raise TypeError(f"{self.__class__.__name__} only supports IrrepsArray")
 
         noises = []
-        out_list = []
-        for (mul, ir), a in zip(x.irreps, x.list):
+        zero_flags = []
+        for (mul, ir), a in zip(x.irreps, x.chunks):
             if self.p >= 1:
-                out_list.append(None)
+                zero_flags.append(True)
                 noises.append(jnp.zeros((mul * ir.dim,), x.dtype))
             elif self.p <= 0:
-                out_list.append(a)
+                zero_flags.append(False)
                 noises.append(jnp.ones((mul * ir.dim,), x.dtype))
+            elif a is None:
+                zero_flags.append(True)
+                noises.append(jnp.zeros((mul * ir.dim,), x.dtype))
             else:
                 noise = jax.random.bernoulli(rng, p=1 - self.p, shape=(mul, 1)) / (
                     1 - self.p
                 )
-                out_list.append(noise * a)
+                zero_flags.append(False)
                 noises.append(jnp.repeat(noise, ir.dim, axis=1).flatten())
 
         noises = jnp.concatenate(noises)
-        return IrrepsArray(irreps=x.irreps, array=x.array * noises, list=out_list)
+        return IrrepsArray(x.irreps, x.array * noises, zero_flags=zero_flags)