+++ e3nn_jax/_src/fc_tp_haiku.py
@@ -16,12 +16,12 @@ class FullyConnectedTensorProduct(hk.Module):
         self, x1: e3nn.IrrepsArray, x2: e3nn.IrrepsArray, **kwargs
     ) -> e3nn.IrrepsArray:
         if self.irreps_in1 is not None:
-            x1 = x1._convert(self.irreps_in1)
+            x1 = x1.rechunk(self.irreps_in1)
         if self.irreps_in2 is not None:
-            x2 = x2._convert(self.irreps_in2)
+            x2 = x2.rechunk(self.irreps_in2)
 
-        x1 = x1.remove_nones().simplify()
-        x2 = x2.remove_nones().simplify()
+        x1 = x1.remove_zero_chunks().simplify()
+        x2 = x2.remove_zero_chunks().simplify()
 
         tp = e3nn.FunctionalFullyConnectedTensorProduct(
             x1.irreps, x2.irreps, self.irreps_out.simplify()
@@ -41,4 +41,4 @@ class FullyConnectedTensorProduct(hk.Module):
             lambda x1, x2: tp.left_right(ws, x1, x2, **kwargs)
         )
         output = f(x1, x2)
-        return output._convert(self.irreps_out)
+        return output.rechunk(self.irreps_out)