+++ e3nn_jax/_src/linear_haiku.py
@@ -149,7 +149,7 @@ class Linear(hk.Module):
             input: e3nn.IrrepsArray = input_or_none
         del weights_or_input, input_or_none
 
-        input = e3nn.IrrepsArray.as_irreps_array(input)
+        input = e3nn.as_irreps_array(input)
 
         dtype = get_pytree_dtype(weights, input)
         if dtype.kind == "i":
@@ -162,7 +162,7 @@ class Linear(hk.Module):
                     f"e3nn.haiku.Linear: The input irreps ({input.irreps}) do not match the expected irreps ({self.irreps_in})"
                 )
 
-        input = input.remove_nones().regroup()
+        input = input.remove_zero_chunks().regroup()
         output_irreps = self.irreps_out.simplify()
 
         if self.channel_out is not None:
@@ -220,4 +220,4 @@ class Linear(hk.Module):
 
         if self.channel_out is not None:
             output = output.mul_to_axis(self.channel_out)
-        return output._convert(self.irreps_out)
+        return output.rechunk(self.irreps_out)