+++ e3nn_jax/_src/symmetric_tensor_product_haiku.py
@@ -93,7 +93,7 @@ class SymmetricTensorProduct(hk.Module):
                 #       out
 
                 if order in self.orders:
-                    for (mul, ir_out), u in zip(U.irreps, U.list):
+                    for (mul, ir_out), u in zip(U.irreps, U.chunks):
                         # u: ndarray [(irreps_x.dim)^order, multiplicity, ir_out.dim]
                         u = (
                             u / u.shape[-2]
@@ -134,7 +134,7 @@ class SymmetricTensorProduct(hk.Module):
 
             # out[irrep_out] : [num_channel, ir_out.dim]
             irreps_out = e3nn.Irreps(sorted(out.keys()))
-            return e3nn.IrrepsArray.from_list(
+            return e3nn.from_chunks(
                 irreps_out,
                 [out[ir][:, None, :] for (_, ir) in irreps_out],
                 (x.shape[0],),